Enhance department, level, task, and course controllers and routes with additional logic

+++ controllers/noticeCtrl.js
@@ -64,6 +64,9 @@ const noticeController = {
     }),
     deleteNotice: asyncHandler(async(req,res)=>{
         const {noticeId} = req.params;
+        if(req.user.role !== admin){
+            throw new Error("You are not authorized to delete this notice")
+        }
         const notice = await Notice.findByIdAndDelete(noticeId);
         if(!notice){
             res.status(400)
